refactor: convert mod into client mod and replace HudMixin with HudRenderCallback
